Fix check for invalid relation references
